refactor: start conversion of TRS client to autogen/wrapper
Use same set-up as WES client (in 5153484), with TRS API client
auto-generated from GA4GH spec and wrapped with methods in the
'TRS' class. Still need to update all methods besides
'get_metadata'.

See #27
