+++ synorchestrator/trs/__init__.py
@@ -1,0 +1,3 @@
+from .client import load_trs_client
+from .wrapper import TRS
+
